Added tests. All tests passed. Build success.

+++ module1/src/main/java/net/broscorp/override/overload/MathOperation.java
@@ -3,7 +3,9 @@ package net.broscorp.override.overload;
 public interface MathOperation {
 
   double someOperation(double var1, double var2);
+
   double someOperation(int var1, int var2);
-  double someOperation(Integer var1, Double var2);
+
+  double someOperation(Integer var1, Integer var2);
 
 }
